Fix missing newlines in podman logs
podman uses almost the same log format as docker, but it strips the '\n'
from any lines it sends. This is fixed by adding a small wrapper that
inserts newline characters at the expected points in the output pipes.

This commit also adds a test case that should ensure logs are working as
expected.