Normalize cancellation behaviour in Wait()
Older docker client versions sometimes generated a "0" status code
on context cancel in ContainerWait(). This then raced the error value
sent on the error channel. A user could interpret this as
"container exit successful" when actually it ran into a timeout.

This commit updates the docker client and normalizes the .Wait()
call to only send something on status channel if the container
actually stopped.