harmonize stop timeouts between docker and podman
When a user calls .Stop() we want to have a similiar timespan until
containers are stopped in both docker and podman. Note that they can't
be exactly the same, as podman does not support sub-second timespans.

This commit sets the default stop timeout to 0 for both docker and
podman. This means that:
* If a timeout of 0 is passed to .Stop(), containers will be killed immediatly
* Podman only: any timeout < 1 second, the containers will be killed immediatly

This makes podman a bit more "eager" when killing containers. This approach was
chosen instead of extending the timeout so that the user expectation of
"Stop() will finish in approx. <= timeout" is preserved.